requirements.txt
black
Django
djangorestframework
jsonfield
psycopg2
psycopg2-binary
requests